Implementation of data model

+++ src/main/java/com/wellsfargo/counselor/entity/Advisor.java
@@ -41,7 +41,7 @@ public class Advisor {
     }
 
     public Long getAdvisorId() {
-        return advisorId;
+        return (Long) advisorId;
     }
 
     public String getFirstName() {
